Adjust Meteor rectify.py and Re-Name/Re-Locate
Update syntax/formatting of python script and re-name/re-locate it to
the image_processors/ directory where it should belong in this new
structure.
